+++ CHANGELOG.md
@@ -6,6 +6,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2022-01-24
+
+### Added
+- Added hierarchical Twig templates. `base.html.twig` acts as the parent template for all other Twig templates and gets selectively overwritten.
+- Added a static HTML view for demonstration purposes.
+- Added Twig debug extension for easier template debugging.
+- Added hint where to start a session and included the session superglobal into the Twig templates if present.
+
+### Changed
+- Updated to fhooe/router v0.3.0.
+- Bumped phpstan/phpstan to 1.4.
+- The result for "GET /form" (PHP form example) is now shown in the same view as "POST /form"
+- The Twig form example is now split up in "GET /twigform" and "POST /twigformresult".
+- Main page is now displayed using Twig.
+- 404 page is now displayed using Twig.
+- Base path is not set automatically. The respective line is commented out so it can easily be included.
+- The 404 handler is now set with an arrow function to demonstrate their usage as well as the automatic inclusion of variables from the parent scope.
+
+### Removed
+- Removed the debugging block that enabled PHP errors. This should be done the webserver configuration.
+
 ## [0.2.0] - 2021-12-22
 
 ### Added
@@ -15,13 +36,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a view for "POST /formresult" using Bootstrap. See also: [3: Use Bootstrap for enhanced views](https://github.com/Digital-Media/fhooe-router-skeleton/issues/3).
 
 ### Changed
-
 - Updated to fhooe/router v0.2.0.
 - Updated the view for "GET /" using Bootstrap. See also: [3: Use Bootstrap for enhanced views](https://github.com/Digital-Media/fhooe-router-skeleton/issues/3).
 - Updated the view for "GET /form" using Bootstrap. See also: [3: Use Bootstrap for enhanced views](https://github.com/Digital-Media/fhooe-router-skeleton/issues/3).
 - Updated the 404 view using Bootstrap. See also [3: Use Bootstrap for enhanced views](https://github.com/Digital-Media/fhooe-router-skeleton/issues/3).
 
 ## [0.1.0] - 2021-12-16
+
 ### Added
 - Added an example invocation of the router in `index.php`.
 - Added three example views for the main page, a form submission and a 404 page.
@@ -32,6 +53,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added notes on Contributing.
 - Added this changelog.
 
-[Unreleased]: https://github.com/Digital-Media/fhooe-router-skeleton/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/Digital-Media/fhooe-router-skeleton/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/Digital-Media/fhooe-router-skeleton/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/Digital-Media/fhooe-router-skeleton/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/Digital-Media/fhooe-router-skeleton/releases/tag/v0.1.0